Refactor ocr default error replacement settings
