+++ ui/controller/ocr_default_error_replacement_settings_controller.py
@@ -1,0 +1,30 @@
+from typing import Dict
+
+from ui.components.ocr_default_error_replacement.ocr_default_error_replacement_settings import (
+    OcrDefaultErrorReplacementSettings,
+)
+from utils.console import console
+
+
+class OcrDefaultErrorReplacementSettingsController:
+    def __init__(self):
+        self.settings = OcrDefaultErrorReplacementSettings(
+            edit_callback=self.edit_button_clicked,
+            delete_callback=self.delete_button_clicked,
+            edit_back_callback=self.edit_back_button_clicked,
+            edit_save_callback=self.edit_save_button_clicked,
+        )
+
+    def edit_button_clicked(self, replacement_map: Dict):
+        self.settings.edit.set_replacement_map(replacement_map)
+        self.settings.layout.setCurrentIndex(1)
+
+    def delete_button_clicked(self, replacement_map: Dict):
+        console.log(replacement_map)
+
+    def edit_back_button_clicked(self):
+        self.settings.layout.setCurrentIndex(0)
+
+    def edit_save_button_clicked(self, replacement_map: Dict):
+        console.log(replacement_map)
+        self.settings.layout.setCurrentIndex(0)
